@@ -5,23 +5,22 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Models.ViewModels.Api;
 
-namespace Api.Attributes
+namespace Api.Attributes;
+
+public class ExceptionFilterAttribute : IExceptionFilter
 {
-    public class ExceptionFilterAttribute : IExceptionFilter
+    private static IEnumerable<string> ResolveExceptionMessages(Exception exception)
     {
-        private static IEnumerable<string> ResolveExceptionMessages(Exception exception)
-        {
-            return exception == null
-                ? Enumerable.Empty<string>()
-                : new[] {exception.Message}.Concat(ResolveExceptionMessages(exception.InnerException));
-        }
+        return exception == null
+            ? []
+            : new[] { exception.Message }.Concat(ResolveExceptionMessages(exception.InnerException));
+    }
 
-        public void OnException(ExceptionContext context)
-        {
-            var result =
-                new BadRequestObjectResult(new ErrorViewModel(ResolveExceptionMessages(context.Exception).ToArray()));
+    public void OnException(ExceptionContext context)
+    {
+        var result =
+            new BadRequestObjectResult(new ErrorViewModel(ResolveExceptionMessages(context.Exception).ToArray()));
 
-            context.Result = result;
-        }
+        context.Result = result;
     }
 }
