@@ -1,26 +1,21 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Api.Controllers
+namespace Api.Controllers;
+
+[ApiExplorerSettings(IgnoreApi = true)]
+[AllowAnonymous]
+[Route("health")]
+public class HealthController : Controller
 {
-    [ApiExplorerSettings(IgnoreApi = true)]
-    [AllowAnonymous]
-    [Route("health")]
-    public class HealthController : Controller
+    [HttpGet]
+    [Route("")]
+    public IActionResult Index()
     {
-        /// <summary>
-        /// Home page
-        /// </summary>
-        /// <returns></returns>
-        [HttpGet]
-        [Route("")]
-        public IActionResult Index()
+        return Ok(new
         {
-            return Ok(new
-            {
-                RemoteIpAddress = HttpContext.Connection.RemoteIpAddress.ToString(),
-                LocalIpAddress = HttpContext.Connection.LocalIpAddress.ToString(),
-            });
-        }
+            RemoteIpAddress = HttpContext.Connection.RemoteIpAddress!.ToString(),
+            LocalIpAddress = HttpContext.Connection.LocalIpAddress!.ToString(),
+        });
     }
 }
