@@ -1,51 +1,23 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Design;
-using Microsoft.Extensions.Configuration;
 using Models.Models;
-using static Dal.Utilities.ConnectionStringUtility;
 
-namespace Dal
+namespace Dal;
+
+public sealed class EntityDbContext: IdentityDbContext<User, IdentityRole<int>, int>
 {
-    public sealed class EntityDbContext: IdentityDbContext<User, IdentityRole<int>, int>, IDesignTimeDbContextFactory<EntityDbContext>
+    public EntityDbContext() { }
+
+    public EntityDbContext(DbContextOptions<EntityDbContext> optionsBuilderOptions) : base(optionsBuilderOptions)
     {
-        public EntityDbContext() { }
+        Database.EnsureCreated();
+    }
 
-        /// <inheritdoc />
-        /// <summary>
-        /// Constructor that will be called by startup.cs
-        /// </summary>
-        /// <param name="optionsBuilderOptions"></param>
-        // ReSharper disable once SuggestBaseTypeForParameter
-        public EntityDbContext(DbContextOptions<EntityDbContext> optionsBuilderOptions) : base(optionsBuilderOptions)
-        {
-            Database.EnsureCreated();
-        }
-
-        protected override void OnModelCreating(ModelBuilder builder)
-        {
-            base.OnModelCreating(builder);
+    protected override void OnModelCreating(ModelBuilder builder)
+    {
+        base.OnModelCreating(builder);
             
-            builder.ApplyConfigurationsFromAssembly(typeof(EntityDbContext).Assembly);
-        }
-
-        /// <summary>
-        ///     This is used for DB migration locally
-        /// </summary>
-        /// <param name="args"></param>
-        /// <returns></returns>
-        public EntityDbContext CreateDbContext(string[] args)
-        {
-            var configuration = new ConfigurationBuilder()
-                .AddEnvironmentVariables()
-                .Build();
-
-            var options = new DbContextOptionsBuilder<EntityDbContext>()
-                .UseNpgsql(ConnectionStringUrlToPgResource(configuration.GetValue<string>("DATABASE_URL")))
-                .Options;
-
-            return new EntityDbContext(options);
-        }
+        builder.ApplyConfigurationsFromAssembly(typeof(EntityDbContext).Assembly);
     }
 }
