@@ -1,13 +1,12 @@
 using System.Net;
 
-namespace Dal.Extensions
+namespace Dal.Extensions;
+
+public static class HttpStatusCodeExtension
 {
-    public static class HttpStatusCodeExtension
+    public static bool IsSuccessStatusCode(this HttpStatusCode statusCode)
     {
-        public static bool IsSuccessStatusCode(this HttpStatusCode statusCode)
-        {
-            var asInt = (int)statusCode;
-            return asInt >= 200 && asInt <= 299;
-        }
+        var asInt = (int)statusCode;
+        return asInt >= 200 && asInt <= 299;
     }
 }
