@@ -1,62 +1,33 @@
 ﻿using System;
 using Npgsql;
-using StackExchange.Redis;
 
-namespace Dal.Utilities
+namespace Dal.Utilities;
+
+public static class ConnectionStringUtility
 {
-    public static class ConnectionStringUtility
+    public static string ConnectionStringUrlToPgResource(string connectionStringUrl)
     {
-        public static string ConnectionStringUrlToRedisResource(string connectionStringUrl)
+        var result = Uri.TryCreate(connectionStringUrl, UriKind.Absolute, out var uri);
+
+        if (!result)
         {
-            var result = Uri.TryCreate(connectionStringUrl, UriKind.Absolute, out var uri);
-
-            if (!result)
-            {
-                throw new ArgumentException("Failed to parse Redis connection string");
-            }
-
-            var userInfo = uri.UserInfo.Split(':');
-            var configurationOptions = new ConfigurationOptions
-            {
-                EndPoints = { { uri.Host, uri.Port } },
-                ClientName = userInfo[0],
-                Password = userInfo[1],
-                AbortOnConnectFail = false
-            };
-
-            return configurationOptions.ToString();
+            throw new ArgumentException("Failed to parse postgres connection string");
         }
 
-        /// <summary>
-        /// Converts connection string url to resource
-        /// </summary>
-        /// <param name="connectionStringUrl"></param>
-        /// <returns></returns>
-        public static string ConnectionStringUrlToPgResource(string connectionStringUrl)
+        var userInfo = uri.UserInfo.Split(':');
+        var connectionStringBuilder = new NpgsqlConnectionStringBuilder
         {
-            var result = Uri.TryCreate(connectionStringUrl, UriKind.Absolute, out var uri);
+            Host = uri.Host,
+            Username = userInfo[0],
+            Password = userInfo[1],
+            Database = uri.Segments[1],
+            ApplicationName = Environment.MachineName,
+            SslMode = SslMode.Require,
+            Pooling = true,
+            // Hard limit
+            MaxPoolSize = 5
+        };
 
-            if (!result)
-            {
-                throw new ArgumentException("Failed to parse postgres connection string");
-            }
-
-            var userInfo = uri.UserInfo.Split(':');
-            var connectionStringBuilder = new NpgsqlConnectionStringBuilder
-            {
-                Host = uri.Host,
-                Username = userInfo[0],
-                Password = userInfo[1],
-                Database = uri.Segments[1],
-                ApplicationName = Environment.MachineName,
-                SslMode = SslMode.Require,
-                TrustServerCertificate = true,
-                Pooling = true,
-                // Hard limit
-                MaxPoolSize = 5
-            };
-
-            return connectionStringBuilder.ToString();
-        }
+        return connectionStringBuilder.ToString();
     }
 }
