@@ -3,28 +3,14 @@ using Logic.Abstracts;
 using Logic.Interfaces;
 using Models.Models;
 
-namespace Logic.Crud
+namespace Logic.Crud;
+
+public class UserLogic(IEfRepository repository) : BasicLogicAbstract<User>, IUserLogic
 {
-    public class UserLogic : BasicLogicAbstract<User>, IUserLogic
+    private readonly IBasicCrud<User> _userDal = repository.For<User>();
+
+    protected override IBasicCrud<User> GetBasicCrudDal()
     {
-        private readonly IBasicCrud<User> _userDal;
-
-        /// <summary>
-        /// Constructor dependency injection
-        /// </summary>
-        /// <param name="repository"></param>
-        public UserLogic(IEfRepository repository)
-        {
-            _userDal = repository.For<User>();
-        }
-
-        /// <summary>
-        /// Returns DAL
-        /// </summary>
-        /// <returns></returns>
-        protected override IBasicCrud<User> GetBasicCrudDal()
-        {
-            return _userDal;
-        }
+        return _userDal;
     }
 }
