@@ -2,20 +2,19 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Logic.Interfaces
+namespace Logic.Interfaces;
+
+public interface IBasicLogic<T>
 {
-    public interface IBasicLogic<T>
-    {
-        Task<IEnumerable<T>> GetAll();
+    Task<IEnumerable<T>> GetAll();
 
-        Task<T> Get(int id);
+    Task<T> Get(int id);
 
-        Task<T> Save(T instance);
+    Task<T> Save(T instance);
         
-        Task<T> Delete(int id);
+    Task<T> Delete(int id);
 
-        Task<T> Update(int id, T dto);
+    Task<T> Update(int id, T dto);
 
-        Task<T> Update(int id, Action<T> updater);
-    }
+    Task<T> Update(int id, Action<T> updater);
 }
