@@ -2,14 +2,13 @@ using System;
 using System.Threading.Tasks;
 using Models.ViewModels.Api;
 
-namespace Logic.Interfaces
+namespace Logic.Interfaces;
+
+public interface IImageUploadLogic
 {
-    public interface IImageUploadLogic
-    {
-        Task<string> Upload(UploadViewModel file);
+    Task<string> Upload(UploadViewModel file);
 
-        Task<UploadViewModel> Download(Guid id);
+    Task<UploadViewModel> Download(Guid id);
 
-        Task<bool> Delete(Guid id);
-    }
+    Task<bool> Delete(Guid id);
 }
