@@ -7,51 +7,43 @@ using Logic.Interfaces;
 using Microsoft.AspNetCore.Http.Internal;
 using Models.ViewModels.Api;
 
-namespace Logic.Logic
+namespace Logic.Logic;
+
+public class ImageUploadLogic(IFileService fileService) : IImageUploadLogic
 {
-    public class ImageUploadLogic : IImageUploadLogic
+    public async Task<string> Upload(UploadViewModel file)
     {
-        private readonly IFileService _fileService;
+        // Randomly assign a key!
+        var key = Guid.NewGuid().ToString();
 
-        public ImageUploadLogic(IFileService fileService)
+        await fileService.Upload(key, file.File.Name, file.File.ContentType, file.File.OpenReadStream(), new Dictionary<string, string>
         {
-            _fileService = fileService;
-        }
-        
-        public async Task<string> Upload(UploadViewModel file)
+            ["Description"] = file.Description
+        });
+
+        return key;
+    }
+
+    public async Task<UploadViewModel> Download(Guid id)
+    {
+        var response = await fileService.Download(id.ToString());
+
+        var formFile = new FormFile(response.Data, 0, response.Data.Length, id.ToString(), response.Name)
         {
-            // Randomly assign a key!
-            var key = Guid.NewGuid().ToString();
+            ContentType = response.ContentType
+        };
 
-            await _fileService.Upload(key, file.File.Name, file.File.ContentType, file.File.OpenReadStream(), new Dictionary<string, string>
-            {
-                ["Description"] = file.Description
-            });
-
-            return key;
-        }
-
-        public async Task<UploadViewModel> Download(Guid id)
+        return new UploadViewModel
         {
-            var response = await _fileService.Download(id.ToString());
+            File = formFile,
+            Description = response.MetaData["Description"]
+        };
+    }
 
-            var formFile = new FormFile(response.Data, 0, response.Data.Length, id.ToString(), response.Name)
-            {
-                ContentType = response.ContentType
-            };
+    public async Task<bool> Delete(Guid id)
+    {
+        var response = await fileService.Delete(id.ToString());
 
-            return new UploadViewModel
-            {
-                File = formFile,
-                Description = response.MetaData["Description"]
-            };
-        }
-
-        public async Task<bool> Delete(Guid id)
-        {
-            var response = await _fileService.Delete(id.ToString());
-
-            return response.Status != HttpStatusCode.BadRequest;
-        }
+        return response.Status != HttpStatusCode.BadRequest;
     }
 }
