@@ -1,22 +1,10 @@
-using System.Collections.Immutable;
-using Models.ViewModels.Config;
+using System.ComponentModel.DataAnnotations;
 
-namespace Models.Constants
+namespace Models.Constants;
+
+public class GlobalConfigs
 {
-    public static class GlobalConfigs
-    {
-        public static ImmutableHashSet<int> StartedStreams { get; set; } = ImmutableHashSet<int>.Empty;
-
-        public static void UpdateGlobalConfigs(GlobalConfigViewModel globalConfigViewModel)
-        {
-        }
-
-        public static GlobalConfigViewModel ToViewModel()
-        {
-            return new GlobalConfigViewModel
-            {
-
-            };
-        }
-    }
+    // We should only have single row of global id in database
+    [Key]
+    public int Id { get; set; }
 }
