@@ -4,24 +4,20 @@ using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.AspNetCore.Identity;
 using Models.Enums;
 
-namespace Models.Models
+namespace Models.Models;
+
+public class User : IdentityUser<int>
 {
-    /// <summary>
-    /// Website user model
-    /// </summary>
-    public class User : IdentityUser<int>
-    {
-        public string Name { get; set; }
+    public string Name { get; set; }
         
-        public virtual DateTimeOffset LastLoginTime { get; set; }  = DateTimeOffset.MinValue;
+    public virtual DateTimeOffset LastLoginTime { get; set; }
         
-        public RoleEnum Role { get; set; }
+    public RoleEnum Role { get; set; }
         
-        [Column(TypeName = "text")]
-        public string Description { get; set; }
+    [Column(TypeName = "text")]
+    public string Description { get; set; }
 
-        public List<Blog> Blogs { get; set; } = new List<Blog>();
+    public List<Unit> Blogs { get; set; }
 
-        public string Photo { get; set; }
-    }
+    public string Photo { get; set; }
 }
