@@ -1,15 +1,9 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Models.ViewModels.Api
-{
-    public class ErrorViewModel
-    {
-        public List<string> Errors { get; }
+namespace Models.ViewModels.Api;
 
-        public ErrorViewModel(params string[] errors)
-        {
-            Errors = errors.Take(1).ToList();
-        }
-    }
+public class ErrorViewModel(params string[] errors)
+{
+    public List<string> Errors { get; } = errors.Take(1).ToList();
 }
