@@ -1,17 +1,16 @@
 using Models.Enums;
 
-namespace Models.ViewModels.Api
+namespace Models.ViewModels.Api;
+
+public class ProfileViewModel
 {
-    public class ProfileViewModel
-    {
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public string Email { get; set; }
+    public string Email { get; set; }
 
-        public string PhoneNumber { get; set; }
+    public string PhoneNumber { get; set; }
 
-        public string Description { get; set; }
+    public string Description { get; set; }
         
-        public RoleEnum Role { get; set; }
-    }
+    public RoleEnum Role { get; set; }
 }
