@@ -1,11 +1,10 @@
 using Microsoft.AspNetCore.Http;
 
-namespace Models.ViewModels.Api
-{
-    public class UploadViewModel
-    {
-        public string Description { get; set; }
+namespace Models.ViewModels.Api;
 
-        public IFormFile File { get; set; }
-    }
+public class UploadViewModel
+{
+    public string Description { get; set; }
+
+    public IFormFile File { get; set; }
 }
