@@ -1,14 +1,13 @@
-namespace Models.ViewModels.Config
+namespace Models.ViewModels.Config;
+
+public class GlobalConfigViewModel
 {
-    public class GlobalConfigViewModel
+    public GlobalConfigViewModel()
     {
-        public GlobalConfigViewModel()
-        {
 
-        }
+    }
 
-        public GlobalConfigViewModel(GlobalConfigViewModel globalConfigViewModel) : this()
-        {
-        }
+    public GlobalConfigViewModel(GlobalConfigViewModel globalConfigViewModel) : this()
+    {
     }
 }
