@@ -1,12 +1,8 @@
-namespace Models.ViewModels.Identities
+namespace Models.ViewModels.Identities;
+
+public class LoginViewModel
 {
-    /// <summary>
-    ///     Login view model
-    /// </summary>
-    public class LoginViewModel
-    {
-        public string Username { get; set; }
+    public string Username { get; set; }
         
-        public string Password { get; set; }
-    }
+    public string Password { get; set; }
 }
