@@ -1,16 +1,12 @@
-namespace Models.ViewModels.Identities
-{
-    /// <summary>
-    ///     Register view model
-    /// </summary>
-    public class RegisterViewModel
-    {
-        public string Name { get; set; }
+namespace Models.ViewModels.Identities;
 
-        public string Username { get; set; }
+public class RegisterViewModel
+{
+    public string Name { get; set; }
+
+    public string Username { get; set; }
         
-        public string Password { get; set; }
+    public string Password { get; set; }
         
-        public string Email { get; set; }
-    }
+    public string Email { get; set; }
 }
