@@ -1,21 +1,15 @@
+namespace Api.Attributes;
+
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 
-namespace Api.Attributes
+public class DisallowAuthorizedFilter : IAuthorizationFilter
 {
-    public class DisallowAuthorizedAttribute : TypeFilterAttribute
+    public void OnAuthorization(AuthorizationFilterContext context)
     {
-        public DisallowAuthorizedAttribute(): base(typeof(DisallowAuthorizedFilter)){}
-    }
-
-    public class DisallowAuthorizedFilter : IAuthorizationFilter
-    {
-        public void OnAuthorization(AuthorizationFilterContext context)
+        if (context.HttpContext.User.Identity is { IsAuthenticated: true })
         {
-            if (context.HttpContext.User.Identity.IsAuthenticated)
-            {
-                context.Result = new BadRequestObjectResult("Authenticated user is disallowed from accessing this route");
-            }
+            context.Result = new BadRequestObjectResult("Authenticated user is disallowed from accessing this route");
         }
     }
 }
