@@ -1,4 +1,6 @@
-﻿using System;
+﻿namespace Api;
+
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
@@ -13,21 +15,13 @@ using Dal;
 using Dal.Configs;
 using Dal.Interfaces;
 using Dal.ServiceApi;
-using EFCache;
-using EFCache.Redis;
 using EfCoreRepository.Extensions;
-using Lamar;
-using LiteDB;
-using Logic.Providers;
-using Logic.Services;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.SignalR;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -37,319 +31,218 @@ using Microsoft.IdentityModel.Logging;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
 using MlkPwgen;
-using Models.Constants;
 using Models.Models;
 using Newtonsoft.Json;
-using StackExchange.Redis;
+using Scrutor;
 using static Dal.Utilities.ConnectionStringUtility;
 
-namespace Api
+public class Startup
 {
-    public class Startup
+    private readonly IWebHostEnvironment _env;
+
+    private readonly IConfigurationRoot _configuration;
+
+    public Startup(IWebHostEnvironment env)
     {
-        private readonly IWebHostEnvironment _env;
+        _env = env;
 
-        private readonly IConfigurationRoot _configuration;
+        var builder = new ConfigurationBuilder()
+            .SetBasePath(env.ContentRootPath)
+            .AddJsonFile("appsettings.json", true, true)
+            .AddJsonFile($"appsettings.{env.EnvironmentName}.json", true)
+            .AddEnvironmentVariables();
 
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="env"></param>
-        public Startup(IWebHostEnvironment env)
+        _configuration = builder.Build();
+    }
+
+    public void ConfigureServices(IServiceCollection services)
+    {
+        services.AddHttpsRedirection(options => options.HttpsPort = 443);
+
+        // If environment is localhost, then enable CORS policy, otherwise no cross-origin access
+        services.AddCors(options => options.AddPolicy("CorsPolicy", builder => builder
+            .WithOrigins(_configuration.GetSection("TrustedSpaUrls").Get<string[]>())
+            .AllowAnyHeader()
+            .AllowAnyMethod()
+            .AllowCredentials()));
+
+        // Add framework services
+        // Add functionality to inject IOptions<T>
+        services.AddOptions();
+
+        services.AddResponseCompression();
+
+        services.Configure<JwtSettings>(_configuration.GetSection("JwtSettings"));
+
+        services.AddLogging();
+
+        services.AddRouting(options => options.LowercaseUrls = true);
+
+        services.AddDistributedMemoryCache();
+
+        services.AddSwaggerGen(config =>
         {
-            _env = env;
-
-            var builder = new ConfigurationBuilder()
-                .SetBasePath(env.ContentRootPath)
-                .AddJsonFile("appsettings.json", true, true)
-                .AddJsonFile($"appsettings.{env.EnvironmentName}.json", true)
-                .AddEnvironmentVariables();
-
-            _configuration = builder.Build();
-        }
-
-        /// <summary>
-        /// This method gets called by the runtime. Use this method to add services to the container.
-        /// For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
-        /// </summary>
-        /// <param name="services"></param>
-        /// <returns></returns>
-        public void ConfigureContainer(ServiceRegistry services)
-        {
-            services.AddHttpsRedirection(options => options.HttpsPort = 443);
-
-            // If environment is localhost, then enable CORS policy, otherwise no cross-origin access
-            services.AddCors(options => options.AddPolicy("CorsPolicy", builder => builder
-                .WithOrigins(_configuration.GetSection("TrustedSpaUrls").Get<string[]>())
-                .AllowAnyHeader()
-                .AllowAnyMethod()
-                .AllowCredentials()));
-
-            // Add framework services
-            // Add functionality to inject IOptions<T>
-            services.AddOptions();
-
-            services.AddResponseCompression();
-
-            services.Configure<JwtSettings>(_configuration.GetSection("JwtSettings"));
-
-            services.AddLogging();
-
-            services.AddRouting(options => options.LowercaseUrls = true);
-
-            if (_env.IsDevelopment())
+            config.SwaggerDoc("v1", new OpenApiInfo
             {
-                services.AddDistributedMemoryCache();
-            }
-            else
-            {
-                var redisConnectionString =
-                    ConnectionStringUrlToRedisResource(_configuration.GetValue<string>("REDISTOGO_URL"));
-
-                services.AddStackExchangeRedisCache(c => c.Configuration = redisConnectionString);
-            }
-
-            services.AddSession(options =>
-            {
-                // Set a short timeout for easy testing.
-                options.IdleTimeout = TimeSpan.FromMinutes(60);
-                options.Cookie.HttpOnly = true;
-                options.Cookie.Name = ApiConstants.AuthenticationSessionCookieName;
-                options.Cookie.SecurePolicy = CookieSecurePolicy.None;
+                Title = "dotnet-template", Version = "v1", Description = "Basic .NET Core template", License =
+                    new OpenApiLicense
+                    {
+                        Name = "MIT",
+                        Url = new Uri("https://github.com/amir734jj/dotnet-template.git"),
+                    },
             });
 
-            services.AddSwaggerGen(config =>
+            // Set the comments path for the Swagger JSON and UI.
+            var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
+            var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
+
+            if (File.Exists(xmlPath))
             {
-                config.SwaggerDoc("v1", new OpenApiInfo
+                config.IncludeXmlComments(xmlPath);
+            }
+
+            config.AddSecurityDefinition(
+                "Bearer",
+                new OpenApiSecurityScheme
                 {
-                    Title = "dotnet-template", Version = "v1", Description = "Basic .NET Core template", License =
-                        new OpenApiLicense
-                        {
-                            Name = "MIT",
-                            Url = new Uri("https://github.com/amir734jj/dotnet-template.git")
-                        }
+                    Name = "Authorization",
+                    In = ParameterLocation.Header,
+                    Description = "JWT Authorization header using the Bearer scheme.",
+                    Type = SecuritySchemeType.Http,
+                    Scheme = "bearer",
                 });
 
-                // Set the comments path for the Swagger JSON and UI.
-                var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
-                var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
-
-                if (File.Exists(xmlPath))
+            config.AddSecurityRequirement(new OpenApiSecurityRequirement
+            {
                 {
-                    config.IncludeXmlComments(xmlPath);
-                }
-
-                config.AddSecurityDefinition("Bearer", // Name the security scheme
-                    // We set the scheme type to http since we're using bearer authentication
-                    // The name of the HTTP Authorization scheme to be used in the Authorization header. In this case "bearer".
                     new OpenApiSecurityScheme
                     {
-                        Name = "Authorization",
-                        In = ParameterLocation.Header,
-                        Description = "JWT Authorization header using the Bearer scheme.",
-                        Type = SecuritySchemeType.Http,
-                        Scheme = "bearer"
-                    });
-
-                config.AddSecurityRequirement(new OpenApiSecurityRequirement
-                {
-                    {
-                        new OpenApiSecurityScheme
+                        Reference = new OpenApiReference
                         {
-                            Reference = new OpenApiReference
-                            {
-                                Type = ReferenceType.SecurityScheme,
-                                Id = "Bearer"
-                            }
-                        }, new List<string>()
-                    }
-                });
+                            Type = ReferenceType.SecurityScheme,
+                            Id = "Bearer",
+                        },
+                    }, new List<string>()
+                },
             });
+        });
 
-            services.AddMvc(x =>
-                {
-                    x.ModelValidatorProviders.Clear();
-
-                    // No need to have https
-                    x.RequireHttpsPermanent = false;
-
-                    // Exception filter attribute
-                    x.Filters.Add<ExceptionFilterAttribute>();
-                }).AddNewtonsoftJson(x =>
-                {
-                    x.SerializerSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore;
-                })
-                .AddRazorPagesOptions(x => x.Conventions.ConfigureFilter(new IgnoreAntiforgeryTokenAttribute()));
-
-            services.AddDbContext<EntityDbContext>(opt =>
+        services.AddMvc(x => x.Filters.Add<ExceptionFilterAttribute>()).AddNewtonsoftJson(x =>
             {
-                if (_env.IsDevelopment())
-                {
-                    opt.UseSqlite(_configuration.GetValue<string>("ConnectionStrings:Sqlite"));
-                }
-                else
-                {
-                    var postgresConnectionString =
-                        ConnectionStringUrlToPgResource(_configuration.GetRequiredValue<string>("DATABASE_URL"));
-                    opt.UseNpgsql(postgresConnectionString);
-                }
-            });
+                x.SerializerSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore;
+            })
+            .AddRazorPagesOptions(x => x.Conventions.ConfigureFilter(new IgnoreAntiforgeryTokenAttribute()));
 
-            services.AddIdentity<User, IdentityRole<int>>(x => x.User.RequireUniqueEmail = false)
-                .AddEntityFrameworkStores<EntityDbContext>()
-                .AddRoles<IdentityRole<int>>()
-                .AddDefaultTokenProviders();
+        services.Scan(scan => scan
+            .FromAssemblies(Assembly.Load("API"), Assembly.Load("Logic"), Assembly.Load("DAL"))
+            .AddClasses() // to register
+            .UsingRegistrationStrategy(RegistrationStrategy.Skip) // 2. Define how to handle duplicates
+            .AsImplementedInterfaces() // 2. Specify which services they are registered as
+            .WithTransientLifetime()); // 3. Set the lifetime for the services
 
-            // L2 EF cache
-            if (_env.IsDevelopment())
-            {
-                EntityFrameworkCache.Initialize(new InMemoryCache());
-            }
-            else
-            {
-                var redisConnectionString =
-                    ConnectionStringUrlToRedisResource(_configuration.GetValue<string>("REDISTOGO_URL"));
-
-                var redisConfigurationOptions = ConfigurationOptions.Parse(redisConnectionString);
-
-                // Important
-                redisConfigurationOptions.AbortOnConnectFail = false;
-
-                EntityFrameworkCache.Initialize(new RedisCache(redisConfigurationOptions));
-            }
-
-            var jwtSetting = _configuration
-                .GetSection("JwtSettings")
-                .Get<JwtSettings>();
-
-            if (_env.IsDevelopment() && string.IsNullOrEmpty(jwtSetting.Key))
-            {
-                jwtSetting.Key = PasswordGenerator.Generate(length: 100, allowed: Sets.Alphanumerics);
-
-                IdentityModelEventSource.ShowPII = true;
-            }
-
-            services.AddAuthentication(options =>
-                {
-                    options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                    options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-                })
-                .AddJwtBearer(config =>
-                {
-                    config.RequireHttpsMetadata = false;
-                    config.SaveToken = true;
-
-                    config.TokenValidationParameters = new TokenValidationParameters
-                    {
-                        ValidIssuer = jwtSetting.Issuer,
-                        ValidAudience = jwtSetting.Audience,
-                        IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(jwtSetting.Key))
-                    };
-                });
-
-            services.AddSingleton<IUserIdProvider, CustomUserIdProvider>();
-
-            services.AddSignalR(config =>
-            {
-                config.MaximumReceiveMessageSize = 10 * 1024 * 1024; // 10 mega-bytes
-                config.StreamBufferCapacity = 50;
-                config.EnableDetailedErrors = true;
-            });
-
-            services.AddEfRepository<EntityDbContext>(x => x.Profile(Assembly.Load("Dal")));
-
-            services.For<JwtSettings>().Use(jwtSetting).Singleton();
-
-            // If environment is localhost then use mock email service
-            if (_env.IsDevelopment())
-            {
-                services.For<ILiteDatabase>().Use(new LiteDatabase("file.litedb")).Singleton();
-                services.For<IFileService>().Use<LiteDbFileService>();
-            }
-            else
-            {
-                var (accessKeyId, secretAccessKey, url) = (
-                    _configuration.GetRequiredValue<string>("CLOUDCUBE_ACCESS_KEY_ID"),
-                    _configuration.GetRequiredValue<string>("CLOUDCUBE_SECRET_ACCESS_KEY"),
-                    _configuration.GetRequiredValue<string>("CLOUDCUBE_URL")
-                );
-
-                var prefix = new Uri(url).Segments.GetValue(1)?.ToString();
-                const string bucketName = "cloud-cube";
-
-                // Generally bad practice
-                var credentials = new BasicAWSCredentials(accessKeyId, secretAccessKey);
-
-                // Create S3 client
-                services.For<IAmazonS3>().Use(_ => new AmazonS3Client(credentials, RegionEndpoint.USEast1));
-                services.For<S3ServiceConfig>().Use(new S3ServiceConfig(bucketName, prefix));
-
-                services.For<IFileService>().Use(ctx => new S3FileService(
-                    ctx.GetInstance<ILogger<S3FileService>>(),
-                    ctx.GetInstance<IAmazonS3>(),
-                    ctx.GetInstance<S3ServiceConfig>()
-                ));
-            }
-
-            // Register stuff in container, using the StructureMap APIs...
-            services.Scan(_ =>
-            {
-                _.AssemblyContainingType(typeof(Startup));
-                _.Assembly("Api");
-                _.Assembly("Logic");
-                _.Assembly("Dal");
-                _.WithDefaultConventions();
-            });
-        }
-
-        /// <summary>
-        /// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        /// </summary>
-        /// <param name="app"></param>
-        public void Configure(IApplicationBuilder app)
+        services.AddDbContext<EntityDbContext>(opt =>
         {
-            app.UseCors("CorsPolicy");
+            var postgresConnectionString =
+                ConnectionStringUrlToPgResource(_configuration.GetRequiredValue<string>("DATABASE_URL"));
+            opt.UseNpgsql(postgresConnectionString);
+        });
 
-            app.UseResponseCompression();
+        services.AddIdentity<User, IdentityRole<int>>(x => x.User.RequireUniqueEmail = false)
+            .AddEntityFrameworkStores<EntityDbContext>()
+            .AddRoles<IdentityRole<int>>()
+            .AddDefaultTokenProviders();
 
-            if (_env.IsDevelopment())
+        var jwtSetting = _configuration
+            .GetSection("JwtSettings")
+            .Get<JwtSettings>();
+
+        if (_env.IsDevelopment() && string.IsNullOrEmpty(jwtSetting.Key))
+        {
+            jwtSetting.Key = PasswordGenerator.Generate(length: 100, allowed: Sets.Alphanumerics);
+
+            IdentityModelEventSource.ShowPII = true;
+        }
+
+        services.AddAuthentication(options =>
             {
-                app.UseDatabaseErrorPage();
-
-                // Enable middleware to serve generated Swagger as a JSON endpoint.
-                app.UseSwagger();
-
-                // Enable middleware to serve swagger-ui (HTML, JS, CSS, etc.), 
-                // specifying the Swagger JSON endpoint.
-                app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "My API V1"));
-            }
-
-            // Not necessary for this workshop but useful when running behind kubernetes
-            app.UseForwardedHeaders(new ForwardedHeadersOptions
+                options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+                options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+            })
+            .AddJwtBearer(config =>
             {
-                // Read and use headers coming from reverse proxy: X-Forwarded-For X-Forwarded-Proto
-                // This is particularly important so that HttpContent.Request.Scheme will be correct behind a SSL terminating proxy
-                ForwardedHeaders = ForwardedHeaders.XForwardedFor |
-                                   ForwardedHeaders.XForwardedProto
+                config.RequireHttpsMetadata = false;
+                config.SaveToken = true;
+
+                config.TokenValidationParameters = new TokenValidationParameters
+                {
+                    ValidIssuer = jwtSetting.Issuer,
+                    ValidAudience = jwtSetting.Audience,
+                    IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(jwtSetting.Key)),
+                };
             });
 
-            // Use wwwroot folder as default static path
-            app.UseDefaultFiles()
-                .UseStaticFiles()
-                .UseCookiePolicy()
-                .UseSession()
-                .UseRouting()
-                .UseCors("CorsPolicy")
-                .UseAuthentication()
-                .UseAuthorization()
-                .UseEndpoints(endpoints =>
-                {
-                    endpoints.MapDefaultControllerRoute();
+        services.AddEfRepository<EntityDbContext>(x => x.Profile(Assembly.Load("Dal")));
 
-                    endpoints.MapHub<MessageHub>("/hub");
-                });
+        services.AddSingleton(jwtSetting);
 
-            Console.WriteLine("Application Started!");
+        var (accessKeyId, secretAccessKey, url) = (
+            _configuration.GetRequiredValue<string>("CLOUDCUBE_ACCESS_KEY_ID"),
+            _configuration.GetRequiredValue<string>("CLOUDCUBE_SECRET_ACCESS_KEY"),
+            _configuration.GetRequiredValue<string>("CLOUDCUBE_URL"));
+
+        var prefix = new Uri(url).Segments.GetValue(1)?.ToString();
+        const string bucketName = "cloud-cube";
+
+        // Generally bad practice
+        var credentials = new BasicAWSCredentials(accessKeyId, secretAccessKey);
+
+        // Create S3 client
+        services.AddSingleton<IAmazonS3>(_ => new AmazonS3Client(credentials, RegionEndpoint.USEast1));
+        services.AddSingleton(new S3ServiceConfig(bucketName, prefix));
+
+        services.AddSingleton<IFileService>(ctx => new S3FileService(
+            ctx.GetRequiredService<ILogger<S3FileService>>(),
+            ctx.GetRequiredService<IAmazonS3>(),
+            ctx.GetRequiredService<S3ServiceConfig>()));
+    }
+
+    public void Configure(IApplicationBuilder app)
+    {
+        app.UseCors("CorsPolicy");
+
+        app.UseResponseCompression();
+
+        if (_env.IsDevelopment())
+        {
+            // Enable middleware to serve generated Swagger as a JSON endpoint.
+            app.UseSwagger();
+
+            // Enable middleware to serve swagger-ui (HTML, JS, CSS, etc.),
+            // specifying the Swagger JSON endpoint.
+            app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "My API V1"));
         }
+
+        // Not necessary for this workshop but useful when running behind kubernetes
+        app.UseForwardedHeaders(new ForwardedHeadersOptions
+        {
+            // Read and use headers coming from reverse proxy: X-Forwarded-For X-Forwarded-Proto
+            // This is particularly important so that HttpContent.Request.Scheme will be correct behind a SSL terminating proxy
+            ForwardedHeaders = ForwardedHeaders.XForwardedFor |
+                               ForwardedHeaders.XForwardedProto,
+        });
+
+        // Use wwwroot folder as default static path
+        app.UseDefaultFiles()
+            .UseStaticFiles()
+            .UseRouting()
+            .UseCors("CorsPolicy")
+            .UseAuthentication()
+            .UseAuthorization()
+            .UseEndpoints(endpoints =>
+            {
+                endpoints.MapDefaultControllerRoute();
+            });
+
+        Console.WriteLine("Application Started!");
     }
 }
